Simplify parser
